Revert "fix: address test feedback for PRD: Ministry Finance Tracking System …" (#523)
This reverts commit f5f95a9ed2e51d7722c2d1abc070ffe8ae11453e.
