+++ app/Actions/LinkDiscordAccount.php
@@ -3,12 +3,8 @@
 namespace App\Actions;
 
 use App\Enums\DiscordAccountStatus;
-use App\Enums\MinecraftAccountStatus;
 use App\Models\DiscordAccount;
-use App\Models\Role;
 use App\Models\User;
-use App\Notifications\AccountLinkedNotification;
-use App\Services\TicketNotificationService;
 use Illuminate\Support\Carbon;
 use Lorisleiva\Actions\Concerns\AsAction;
 
@@ -60,37 +56,6 @@ class LinkDiscordAccount
             "Linked Discord account: {$account->username} ({$account->discord_user_id})"
         );
 
-        $this->sendAccountLinkedNotifications($user, $account->username, 'Discord');
-
         return ['success' => true, 'message' => 'Discord account linked successfully.', 'account' => $account];
-    }
-
-    private function sendAccountLinkedNotifications(User $user, string $accountName, string $accountType): void
-    {
-        $activeMinecraft = $user->minecraftAccounts()->where('status', MinecraftAccountStatus::Active)->count();
-        $disabledMinecraft = $user->minecraftAccounts()->where('status', '!=', MinecraftAccountStatus::Active->value)->count();
-        $activeDiscord = $user->discordAccounts()->where('status', DiscordAccountStatus::Active)->count();
-        $disabledDiscord = $user->discordAccounts()->where('status', '!=', DiscordAccountStatus::Active->value)->count();
-
-        $notification = new AccountLinkedNotification(
-            $user, $accountName, $accountType,
-            $activeMinecraft, $disabledMinecraft, $activeDiscord, $disabledDiscord
-        );
-
-        $service = app(TicketNotificationService::class);
-
-        // Notify parent (mail only, unconditional)
-        foreach ($user->parents as $parent) {
-            $parent->notify((clone $notification)->setChannels(['mail']));
-        }
-
-        // Notify staff with 'User - Manager' role via their staff_alerts preferences
-        $userManagerRoleId = Role::where('name', 'User - Manager')->value('id');
-        if ($userManagerRoleId) {
-            $managers = User::whereHas('staffPosition.roles', fn ($r) => $r->where('roles.id', $userManagerRoleId))->get();
-            foreach ($managers as $manager) {
-                $service->send($manager, clone $notification, 'staff_alerts');
-            }
-        }
     }
 }
